requirements.txt
asgiref==3.9.1
astroid==3.3.11
autopep8==2.3.2
colorama==0.4.6
dill==0.4.0
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
isort==6.0.1
mccabe==0.7.0
mysqlclient==2.2.7
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pylint==3.3.8
python-decouple==3.8
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2025.2
whitenoise==6.9.0
